adjust file and do some refactoring
