+++ src/contexts/theme.ts
@@ -1,8 +1,8 @@
 "use client";
 
-import { createTheme } from "@mui/material/styles";
 import { ibmPlexSansThai } from "@/config/fonts";
-import { Button as MuiButton } from "@/config/button";
+import { MuiButton } from "@/styles/mui/button";
+import { createTheme } from "@mui/material/styles";
 
 const defaultTheme = createTheme();
 const fontFamily = [
